PID function a little pushy #427

+++ src/homeform.cpp
@@ -3037,7 +3037,9 @@ void homeform::update() {
             if (last_seconds_pid_heart_zone == 0 || ((seconds - last_seconds_pid_heart_zone) >= delta)) {
 
                 last_seconds_pid_heart_zone = seconds;
-                uint8_t zone = settings.value(QStringLiteral("treadmill_pid_heart_zone"), QStringLiteral("Disabled"))
+
+                // it must be double because the current zone is double
+                double zone = settings.value(QStringLiteral("treadmill_pid_heart_zone"), QStringLiteral("Disabled"))
                                    .toString()
                                    .toUInt();
                 if (fromTrainProgram) {
